feat: 홈 추천 이미지 불러오기 기능 추가
