+++ app/src/main/java/com/example/kotlin_starbucks/network/NetworkAPI.kt
@@ -1,9 +1,6 @@
 package com.example.kotlin_starbucks.network
 
-import com.example.kotlin_starbucks.model.EventImageContents
-import com.example.kotlin_starbucks.model.HomeProducts
-import com.example.kotlin_starbucks.model.ProductCd
-import com.example.kotlin_starbucks.model.YourRecommendProducts
+import com.example.kotlin_starbucks.model.*
 import retrofit2.Response
 import retrofit2.http.*
 
@@ -19,5 +16,11 @@ interface NetworkAPI {
     @POST("productViewAjax.do")
     suspend fun loadStarbucksContents(
         @Field("product_cd") productCd: Long?
-    ): Response<YourRecommendProducts>
+    ): Response<YourRecommendProductsTitle>
+
+    @FormUrlEncoded
+    @POST("productFileAjax.do")
+    suspend fun loadStarbucksContentsImage(
+        @Field("PRODUCT_CD") productCd: Long?
+    ): Response<YourRecommendProductsImage>
 }
