+++ app/src/main/java/com/example/kotlin_starbucks/ui/HomeAdapter.kt
@@ -6,9 +6,9 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.example.kotlin_starbucks.databinding.ItemYourRecommendBinding
-import com.example.kotlin_starbucks.model.File
+import com.example.kotlin_starbucks.model.YourRecommendProducts
 
-class HomeAdapter : ListAdapter<File, HomeAdapter.HomeHolder>(FileDiffCallBack) {
+class HomeAdapter : ListAdapter<YourRecommendProducts, HomeAdapter.HomeHolder>(FileDiffCallBack) {
 
     private lateinit var binding: ItemYourRecommendBinding
 
@@ -17,30 +17,29 @@ class HomeAdapter : ListAdapter<File, HomeAdapter.HomeHolder>(FileDiffCallBack) 
         return HomeHolder(binding)
     }
 
-
     override fun onBindViewHolder(holder: HomeHolder, position: Int) {
         holder.bind(getItem(position))
     }
 
     class HomeHolder(private val binding: ItemYourRecommendBinding) :
         RecyclerView.ViewHolder(binding.root) {
-        fun bind(file: File) {
-            binding.product = file
+        fun bind(products: YourRecommendProducts) {
+            binding.product = products
             binding.executePendingBindings()
         }
     }
 
-    object FileDiffCallBack : DiffUtil.ItemCallback<File>() {
+    object FileDiffCallBack : DiffUtil.ItemCallback<YourRecommendProducts>() {
         override fun areItemsTheSame(
-            oldItem: File,
-            newItem: File
+            oldItem: YourRecommendProducts,
+            newItem: YourRecommendProducts
         ): Boolean {
             return oldItem.hashCode() == newItem.hashCode()
         }
 
         override fun areContentsTheSame(
-            oldItem: File,
-            newItem: File
+            oldItem: YourRecommendProducts,
+            newItem: YourRecommendProducts
         ): Boolean {
             return oldItem == newItem
         }
