+++ app/src/main/java/com/example/kotlin_starbucks/ui/ViewModel.kt
@@ -1,6 +1,5 @@
 package com.example.kotlin_starbucks.ui
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -21,9 +20,17 @@ class ViewModel @Inject constructor(private val repository: Repository) : ViewMo
     private val _homeContents = MutableLiveData<HomeProducts>()
     val homeContents: LiveData<HomeProducts> = _homeContents
 
-    private val _homeContentsDetailList: MutableList<File> = mutableListOf()
-    private val _homeContentsDetail = MutableLiveData<MutableList<File>>()
-    val homeContentsDetail: LiveData<MutableList<File>> = _homeContentsDetail
+    private val _homeContentsDetailList: MutableList<Details> = mutableListOf()
+    private val _homeContentsDetail = MutableLiveData<MutableList<Details>>()
+    val homeContentsDetail: LiveData<MutableList<Details>> = _homeContentsDetail
+
+    private val _homeContentsDetailImageList: MutableList<String> = mutableListOf()
+    private val _homeContentsDetailImage = MutableLiveData<MutableList<String>>()
+    val homeContentsDetailImage: LiveData<MutableList<String>> = _homeContentsDetailImage
+
+    private val _yourRecommendProductsList: MutableList<YourRecommendProducts> = mutableListOf()
+    private val _yourRecommendProducts = MutableLiveData<MutableList<YourRecommendProducts>>()
+    val yourRecommendProducts: LiveData<MutableList<YourRecommendProducts>> = _yourRecommendProducts
 
     private val _error = SingleLiveEvent<ImageException>()
     val error: LiveData<ImageException> = _error
@@ -56,16 +63,39 @@ class ViewModel @Inject constructor(private val repository: Repository) : ViewMo
                 }
             }
             loadHomeContentsJob.join()
-            for (i in 0 until homeContents.value?.yourRecommend?.products?.size!!) {
-                launch {
-                    val yourRecommendProducts = repository.loadStarbucksContents(homeContents.value?.yourRecommend?.products!![i]?.toLong())
-                    yourRecommendProducts?.view?.let {
-                        _homeContentsDetailList.add(it)
-                        _homeContentsDetail.value = _homeContentsDetailList
+            launch {
+                for (i in 0 until homeContents.value?.yourRecommend?.products?.size!!) {
+                    launch {
+                        val yourRecommendProducts = repository.loadStarbucksContents(homeContents.value?.yourRecommend?.products!![i]?.toLong())
+                        yourRecommendProducts?.view?.let {
+                            _homeContentsDetailList.add(it)
+                            _homeContentsDetail.value = _homeContentsDetailList
+                        }
+                        val recommendProductImage = repository.loadStarbucksImages(homeContents.value?.yourRecommend?.products!![i]?.toLong())
+                        if (!recommendProductImage?.file.isNullOrEmpty()) {
+                            recommendProductImage?.file?.get(0)?.filePATH.let {
+                                if (it != null) {
+                                    _homeContentsDetailImageList.add(it)
+                                    _homeContentsDetailImage.value = _homeContentsDetailImageList
+                                }
+                            }
+                        }
                     }
                 }
-            }
+            }.join()
+            makeProductsList()
         }
+    }
+
+    private fun makeProductsList() {
+        for (index in 0 until _homeContentsDetailList.size) {
+            _yourRecommendProductsList.add(
+                YourRecommendProducts(
+                    _homeContentsDetail.value?.get(index)?.productNM,
+                    _homeContentsDetailImage.value?.get(index)
+            ))
+        }
+        _yourRecommendProducts.value = _yourRecommendProductsList
     }
 
     private fun <T> MutableLiveData<T>.notifyObserver() {
