+++ app/src/main/java/com/example/kotlin_starbucks/ui/bindingAdapter/ImageBindingAdapter.kt
@@ -2,6 +2,7 @@ package com.example.kotlin_starbucks.ui.bindingAdapter
 
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
+import com.example.kotlin_starbucks.network.RetrofitObject
 import com.example.kotlin_starbucks.repository.module.GlideApp
 
 @BindingAdapter("imageUrl")
@@ -9,6 +10,15 @@ fun setEventImageBitmap(view: ImageView, imageUrl: String?) {
     if (!imageUrl.isNullOrEmpty()) {
         GlideApp.with(view)
             .load(imageUrl)
+            .into(view)
+    }
+}
+
+@BindingAdapter("recommendImageUrl")
+fun setRecommendImage(view: ImageView, imageUrl: String?) {
+    if (!imageUrl.isNullOrEmpty()) {
+        GlideApp.with(view)
+            .load(RetrofitObject.baseImageUrl + imageUrl)
             .into(view)
     }
 }
