+++ app/src/main/res/drawable/item_image_shape.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+
+   <!-- <corners android:radius="200dp"/>-->
+
+</shape>
