+++ app/src/main/res/layout/fragment_home.xml
@@ -53,8 +53,7 @@
             android:id="@+id/rv_your_recommend"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:paddingStart="10dp"
-            android:paddingEnd="10dp"
+
             app:layout_constraintTop_toBottomOf="@+id/tv_your_recommend_title"
             app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
             app:layout_constraintStart_toStartOf="parent"
